Fix memory of Metadata Form View
